Beginnings of list support, bunch of cleanup
