+++ tests/list_test.py
@@ -1,0 +1,3 @@
+a_list = []
+for i in range(10):
+    a_list.append(i)
